feat(cloudx): implement login
